requirements.txt
pytest >= 8.2.0
selene >= 2.0.0rc9
allure-pytest >= 2.13.5
selenium
python-dotenv >= 1.0.1
